Move all tests from aicsimage to their own readers
