+++ aicsimageio/tests/readers/extra_readers/test_czi_reader.py
@@ -2,19 +2,22 @@
 # -*- coding: utf-8 -*-
 
 import xml.etree.ElementTree as ET
-from typing import List, Tuple
+from typing import Any, List, Optional, Tuple, Union
+from pathlib import Path
+from tempfile import TemporaryDirectory
 
 import numpy as np
 import pytest
 
-from aicsimageio import dimensions, exceptions
+from aicsimageio import AICSImage, dimensions, exceptions
 from aicsimageio.readers import ArrayLikeReader
 from aicsimageio.readers.czi_reader import CziReader
 
-from ..conftest import LOCAL, REMOTE, get_resource_full_path
-from ..image_container_test_utils import (
+from ...conftest import LOCAL, REMOTE, get_resource_full_path
+from ...image_container_test_utils import (
     run_image_container_mosaic_checks,
     run_image_file_checks,
+    run_multi_scene_image_read_checks,
 )
 
 
@@ -406,3 +409,255 @@ def test_czi_reader_mosaic_coords(
         reader.mosaic_xarray_dask_data.coords[dimensions.DimensionNames.SpatialX].data,
         expected_mosaic_x_coords,
     )
+
+
+@pytest.mark.parametrize(
+    "filename, "
+    "set_scene, "
+    "expected_scenes, "
+    "expected_shape, "
+    "expected_dtype, "
+    "expected_dims_order, "
+    "expected_channel_names, "
+    "expected_physical_pixel_sizes, "
+    "expected_metadata_type",
+    [
+        (
+            "s_1_t_1_c_1_z_1.czi",
+            "Image:0",
+            ("Image:0",),
+            (1, 1, 1, 325, 475),
+            np.uint16,
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["Bright"],
+            (None, 1.0833333333333333, 1.0833333333333333),
+            ET.Element,
+        ),
+        (
+            "RGB-8bit.czi",
+            "Image:0",
+            ("Image:0",),
+            (1, 1, 1, 624, 924, 3),
+            np.uint8,
+            dimensions.DEFAULT_DIMENSION_ORDER_WITH_SAMPLES,
+            ["Channel:0:0"],
+            (None, 1.0833333333333333, 1.0833333333333333),
+            ET.Element,
+        ),
+    ],
+)
+def test_aicsimage(
+    filename: str,
+    set_scene: str,
+    expected_scenes: Tuple[str, ...],
+    expected_shape: Tuple[int, ...],
+    expected_dtype: np.dtype,
+    expected_dims_order: str,
+    expected_channel_names: List[str],
+    expected_physical_pixel_sizes: Tuple[float, float, float],
+    expected_metadata_type: Union[type, Tuple[Union[type, Tuple[Any, ...]], ...]],
+) -> None:
+    # Construct full filepath
+    uri = get_resource_full_path(filename, LOCAL)
+
+    # Run checks
+    run_image_file_checks(
+        ImageContainer=AICSImage,
+        image=uri,
+        set_scene=set_scene,
+        expected_scenes=expected_scenes,
+        expected_current_scene=set_scene,
+        expected_shape=expected_shape,
+        expected_dtype=expected_dtype,
+        expected_dims_order=expected_dims_order,
+        expected_channel_names=expected_channel_names,
+        expected_physical_pixel_sizes=expected_physical_pixel_sizes,
+        expected_metadata_type=expected_metadata_type,
+    )
+
+
+@pytest.mark.parametrize(
+    "filename, "
+    "first_scene_id, "
+    "first_scene_shape, "
+    "second_scene_id, "
+    "second_scene_shape",
+    [
+        (
+            "s_3_t_1_c_3_z_5.czi",
+            "P2",
+            (1, 3, 5, 325, 475),
+            "P3",
+            (1, 3, 5, 325, 475),
+        ),
+        (
+            "s_3_t_1_c_3_z_5.czi",
+            "P3",
+            (1, 3, 5, 325, 475),
+            "P1",
+            (1, 3, 5, 325, 475),
+        ),
+        (
+            "s_3_t_1_c_3_z_5.czi",
+            1,
+            (1, 3, 5, 325, 475),
+            2,
+            (1, 3, 5, 325, 475),
+        ),
+        pytest.param(
+            "s_3_t_1_c_3_z_5.czi",
+            ["this is not a scene id"],
+            None,
+            None,
+            None,
+            marks=pytest.mark.raises(exception=TypeError),
+        ),
+    ],
+)
+def test_multi_scene_aicsimage(
+    filename: str,
+    first_scene_id: str,
+    first_scene_shape: Tuple[int, ...],
+    second_scene_id: str,
+    second_scene_shape: Tuple[int, ...],
+) -> None:
+    # Construct full filepath
+    uri = get_resource_full_path(filename, LOCAL)
+
+    # Run checks
+    run_multi_scene_image_read_checks(
+        ImageContainer=AICSImage,
+        image=uri,
+        first_scene_id=first_scene_id,
+        first_scene_shape=first_scene_shape,
+        first_scene_dtype=np.dtype(np.uint16),
+        second_scene_id=second_scene_id,
+        second_scene_shape=second_scene_shape,
+        second_scene_dtype=np.dtype(np.uint16),
+    )
+
+
+@pytest.mark.parametrize(
+    "filename, expected_shape",
+    [
+        ("RGB-8bit.czi", (1, 1, 1, 624, 924, 3)),
+    ],
+)
+def test_no_scene_prop_access(
+    filename: str,
+    expected_shape: Tuple[int, ...],
+) -> None:
+    # Construct full filepath
+    uri = get_resource_full_path(filename, LOCAL)
+
+    # Construct image and check no scene call with property access
+    img = AICSImage(uri)
+    assert img.shape == expected_shape
+
+
+@pytest.mark.parametrize(
+    "filename, select_scenes",
+    [
+        ("s_1_t_1_c_1_z_1.czi", None),
+        ("s_3_t_1_c_3_z_5.czi", ["P2", "P1"]),
+    ],
+)
+def test_roundtrip_save_all_scenes(
+    filename: str, select_scenes: Optional[List[str]]
+) -> None:
+    # Construct full filepath
+    uri = get_resource_full_path(filename, LOCAL)
+
+    # Read initial
+    original = AICSImage(uri)
+
+    # Save to temp and compare
+    with TemporaryDirectory() as tmpdir:
+        save_path = Path(tmpdir) / f"converted-{filename}.ome.tiff"
+        original.save(save_path, select_scenes=select_scenes)
+
+        # Re-read
+        result = AICSImage(save_path)
+
+        # Compare all scenes
+        # They may not have the same scene ids as some readers use scene names as the
+        # id, see LifReader for example
+        if select_scenes is None:
+            select_original_scenes = list(original.scenes)
+        else:
+            select_original_scenes = select_scenes
+
+        assert len(select_original_scenes) == len(result.scenes)
+        for original_scene_id, result_scene_id in zip(
+            select_original_scenes, result.scenes
+        ):
+            # Compare
+            original.set_scene(original_scene_id)
+            result.set_scene(result_scene_id)
+
+            np.testing.assert_array_equal(original.data, result.data)
+            assert original.dims.order == result.dims.order
+            assert original.channel_names == result.channel_names
+
+
+@pytest.mark.parametrize(
+    "filename, "
+    "reconstruct_mosaic, "
+    "set_scene, "
+    "expected_shape, "
+    "expected_mosaic_tile_dims, "
+    "specific_tile_index",
+    [
+        (
+            "OverViewScan.czi",
+            True,
+            "TR1",
+            (1, 1, 1, 3212, 7398),
+            (440, 544),
+            0,
+        ),
+        (
+            "OverViewScan.czi",
+            False,
+            "TR1",
+            (120, 1, 1, 1, 440, 544),
+            (440, 544),
+            0,
+        ),
+        pytest.param(
+            "actk.ome.tiff",
+            True,
+            "Image:0",
+            (1, 6, 65, 233, 345),
+            None,
+            0,
+            # AttributeError raises not because of error in rollback
+            # but because cannot access Y or X from
+            # None return from `mosaic_tile_dims` because
+            # image is not a mosaic tiled image
+            marks=pytest.mark.raises(exception=AttributeError),
+        ),
+    ],
+)
+def test_mosaic_passthrough(
+    filename: str,
+    reconstruct_mosaic: bool,
+    set_scene: str,
+    expected_shape: Tuple[int, ...],
+    expected_mosaic_tile_dims: Tuple[int, int],
+    specific_tile_index: int,
+) -> None:
+    # Construct full filepath
+    uri = get_resource_full_path(filename, LOCAL)
+
+    # Init
+    img = AICSImage(uri, reconstruct_mosaic=reconstruct_mosaic)
+    img.set_scene(set_scene)
+
+    # Assert basics
+    assert img.shape == expected_shape
+    assert img.mosaic_tile_dims.Y == expected_mosaic_tile_dims[0]  # type: ignore
+    assert img.mosaic_tile_dims.X == expected_mosaic_tile_dims[1]  # type: ignore
+
+    # Ensure that regardless of stitched or not, we can get tile position
+    img.get_mosaic_tile_position(specific_tile_index)
