+++ aicsimageio/tests/readers/extra_readers/test_dv_reader.py
@@ -1,16 +1,16 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from typing import List, Tuple
+from typing import Any, List, Tuple, Union
 
 import numpy as np
 import pytest
 
-from aicsimageio import dimensions, exceptions
+from aicsimageio import AICSImage, dimensions, exceptions
 from aicsimageio.readers import DVReader
 from aicsimageio.tests.image_container_test_utils import run_image_file_checks
 
-from ..conftest import get_resource_full_path, host
+from ...conftest import get_resource_full_path, host, LOCAL
 
 
 @host
@@ -84,4 +84,58 @@ def test_dv_reader(
         expected_channel_names=expected_channel_names,
         expected_physical_pixel_sizes=expected_physical_pixel_sizes,
         expected_metadata_type=dict,
+    )
+
+
+@pytest.mark.parametrize(
+    "filename, "
+    "set_scene, "
+    "expected_scenes, "
+    "expected_shape, "
+    "expected_dtype, "
+    "expected_dims_order, "
+    "expected_channel_names, "
+    "expected_physical_pixel_sizes, "
+    "expected_metadata_type",
+    [
+        (
+            "DV_siRNAi-HeLa_IN_02.r3d_D3D.dv",
+            "Image:0",
+            ("Image:0",),
+            (1, 4, 40, 512, 512),
+            np.int16,
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["360/457", "490/528", "555/617", "640/685"],
+            (0.20000000298023224, 0.06502940505743027, 0.06502940505743027),
+            dict,
+        ),
+    ],
+)
+def test_aicsimage(
+    filename: str,
+    set_scene: str,
+    expected_scenes: Tuple[str, ...],
+    expected_shape: Tuple[int, ...],
+    expected_dtype: np.dtype,
+    expected_dims_order: str,
+    expected_channel_names: List[str],
+    expected_physical_pixel_sizes: Tuple[float, float, float],
+    expected_metadata_type: Union[type, Tuple[Union[type, Tuple[Any, ...]], ...]],
+) -> None:
+    # Construct full filepath
+    uri = get_resource_full_path(filename, LOCAL)
+
+    # Run checks
+    run_image_file_checks(
+        ImageContainer=AICSImage,
+        image=uri,
+        set_scene=set_scene,
+        expected_scenes=expected_scenes,
+        expected_current_scene=set_scene,
+        expected_shape=expected_shape,
+        expected_dtype=expected_dtype,
+        expected_dims_order=expected_dims_order,
+        expected_channel_names=expected_channel_names,
+        expected_physical_pixel_sizes=expected_physical_pixel_sizes,
+        expected_metadata_type=expected_metadata_type,
     )
