+++ aicsimageio/tests/readers/extra_readers/test_nd2_reader.py
@@ -1,16 +1,16 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from typing import List, Tuple
+from typing import Any, List, Tuple, Union
 
 import numpy as np
 import pytest
 
-from aicsimageio import exceptions
+from aicsimageio import AICSImage, dimensions, exceptions
 from aicsimageio.readers.nd2_reader import ND2Reader
 from aicsimageio.tests.image_container_test_utils import run_image_file_checks
 
-from ..conftest import get_resource_full_path, host
+from ...conftest import get_resource_full_path, host, LOCAL
 
 
 @host
@@ -164,4 +164,58 @@ def test_nd2_reader(
         expected_channel_names=expected_channel_names,
         expected_physical_pixel_sizes=expected_physical_pixel_sizes,
         expected_metadata_type=dict,
+    )
+
+
+@pytest.mark.parametrize(
+    "filename, "
+    "set_scene, "
+    "expected_scenes, "
+    "expected_shape, "
+    "expected_dtype, "
+    "expected_dims_order, "
+    "expected_channel_names, "
+    "expected_physical_pixel_sizes, "
+    "expected_metadata_type",
+    [
+        (
+            "ND2_jonas_header_test2.nd2",
+            "XYPos:0",
+            ("XYPos:0",),
+            (4, 1, 5, 520, 696),
+            np.uint16,
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["Jonas_DIC"],
+            (0.5, 0.12863494437945, 0.12863494437945),
+            dict,
+        ),
+    ],
+)
+def test_aicsimage(
+    filename: str,
+    set_scene: str,
+    expected_scenes: Tuple[str, ...],
+    expected_shape: Tuple[int, ...],
+    expected_dtype: np.dtype,
+    expected_dims_order: str,
+    expected_channel_names: List[str],
+    expected_physical_pixel_sizes: Tuple[float, float, float],
+    expected_metadata_type: Union[type, Tuple[Union[type, Tuple[Any, ...]], ...]],
+) -> None:
+    # Construct full filepath
+    uri = get_resource_full_path(filename, LOCAL)
+
+    # Run checks
+    run_image_file_checks(
+        ImageContainer=AICSImage,
+        image=uri,
+        set_scene=set_scene,
+        expected_scenes=expected_scenes,
+        expected_current_scene=set_scene,
+        expected_shape=expected_shape,
+        expected_dtype=expected_dtype,
+        expected_dims_order=expected_dims_order,
+        expected_channel_names=expected_channel_names,
+        expected_physical_pixel_sizes=expected_physical_pixel_sizes,
+        expected_metadata_type=expected_metadata_type,
     )
