+++ aicsimageio/tests/readers/test_array_like_reader.py
@@ -8,7 +8,7 @@ import numpy as np
 import pytest
 import xarray as xr
 
-from aicsimageio import exceptions, types
+from aicsimageio import AICSImage, dimensions, exceptions, types
 from aicsimageio.readers import ArrayLikeReader
 
 from ..image_container_test_utils import run_image_container_checks
@@ -839,6 +839,281 @@ def test_arraylike_reader(
     image_container = ArrayLikeReader(
         image, dim_order=dim_order, channel_names=channel_names
     )
+
+    run_image_container_checks(
+        image_container=image_container,
+        set_scene=set_scene,
+        expected_scenes=expected_scenes,
+        expected_current_scene=set_scene,
+        expected_shape=expected_shape,
+        expected_dtype=np.dtype(np.float64),
+        expected_dims_order=expected_dims,
+        expected_channel_names=expected_channel_names,
+        expected_physical_pixel_sizes=(None, None, None),
+        # we allow both None and Dict because the user can pass an already initialized
+        # xarray DataArray which has metadata as a dict
+        expected_metadata_type=(
+            type(None),
+            dict,
+        ),
+    )
+
+
+@pytest.mark.parametrize(
+    "image, "
+    "dim_order, "
+    "channel_names, "
+    "set_scene, "
+    "expected_scenes, "
+    "expected_shape, "
+    "expected_dims, "
+    "expected_channel_names",
+    [
+        # Check no metadata provided 2D
+        # these are really simple just None checks
+        (
+            np.random.rand(1, 1),
+            None,
+            None,
+            "Image:0",
+            ("Image:0",),
+            (1, 1, 1, 1, 1),
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["Channel:0:0"],
+        ),
+        (
+            xr.DataArray(np.random.rand(1, 1)),
+            None,
+            None,
+            "Image:0",
+            ("Image:0",),
+            (1, 1, 1, 1, 1),
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["Channel:0:0"],
+        ),
+        # Check no metadata provided 4D
+        # these check that channel names are created for all
+        # and specifically for xr that channel names are overwritten
+        (
+            np.random.rand(1, 1, 1, 1),
+            None,
+            None,
+            "Image:0",
+            ("Image:0",),
+            (1, 1, 1, 1, 1),
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["Channel:0:0"],
+        ),
+        (
+            xr.DataArray(np.random.rand(1, 1, 1, 1)),
+            None,
+            None,
+            "Image:0",
+            ("Image:0",),
+            (1, 1, 1, 1, 1),
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["Channel:0:0"],
+        ),
+        # Test many scene, same dim_order, second scene
+        (
+            [np.random.rand(1, 1, 1), np.random.rand(2, 2, 2)],
+            "CYX",
+            None,
+            "Image:1",
+            ("Image:0", "Image:1"),
+            (1, 2, 1, 2, 2),
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["Channel:1:0", "Channel:1:1"],
+        ),
+        (
+            [
+                xr.DataArray(np.random.rand(1, 1, 1)),
+                xr.DataArray(np.random.rand(2, 2, 2)),
+            ],
+            "CYX",
+            None,
+            "Image:1",
+            ("Image:0", "Image:1"),
+            (1, 2, 1, 2, 2),
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["Channel:1:0", "Channel:1:1"],
+        ),
+        # Test many scene, different dim_order, different channel_names, second scene
+        (
+            [np.random.rand(1, 1, 1), np.random.rand(2, 2, 2)],
+            [None, "CYX"],
+            [None, ["A", "B"]],
+            "Image:1",
+            ("Image:0", "Image:1"),
+            (1, 2, 1, 2, 2),
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["A", "B"],
+        ),
+        (
+            [
+                xr.DataArray(np.random.rand(1, 1, 1)),
+                xr.DataArray(np.random.rand(2, 2, 2)),
+            ],
+            [None, "CYX"],
+            [None, ["A", "B"]],
+            "Image:1",
+            ("Image:0", "Image:1"),
+            (1, 2, 1, 2, 2),
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["A", "B"],
+        ),
+        # Test filled in xarray(s)
+        # no metadata should be overwritten
+        (
+            xr.DataArray(np.random.rand(1, 1, 1), dims=list("TYX")),
+            None,
+            None,
+            "Image:0",
+            ("Image:0",),
+            (1, 1, 1, 1, 1),
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["Channel:0:0"],
+        ),
+        (
+            xr.DataArray(
+                np.random.rand(1, 1, 1), dims=list("CYX"), coords={"C": ["A"]}
+            ),
+            None,
+            None,
+            "Image:0",
+            ("Image:0",),
+            (1, 1, 1, 1, 1),
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["A"],
+        ),
+        (
+            [
+                xr.DataArray(np.random.rand(1, 1, 1), dims=list("TYX")),
+                xr.DataArray(
+                    np.random.rand(2, 2, 2), dims=list("CYX"), coords={"C": ["A", "B"]}
+                ),
+            ],
+            None,
+            None,
+            "Image:0",
+            ("Image:0", "Image:1"),
+            (1, 1, 1, 1, 1),
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["Channel:0:0"],
+        ),
+        (
+            [
+                xr.DataArray(np.random.rand(1, 1, 1), dims=list("TYX")),
+                xr.DataArray(
+                    np.random.rand(2, 2, 2), dims=list("CYX"), coords={"C": ["A", "B"]}
+                ),
+            ],
+            None,
+            None,
+            "Image:1",
+            ("Image:0", "Image:1"),
+            (1, 2, 1, 2, 2),
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["A", "B"],
+        ),
+        # Test non-standard dimensions
+        (
+            np.random.rand(2, 2, 2),
+            "ABD",
+            None,
+            "Image:0",
+            ("Image:0",),
+            (1, 1, 1, 1, 1),
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["Channel:0:0"],
+        ),
+        (
+            xr.DataArray(np.random.rand(2, 2, 2)),
+            "ABD",
+            None,
+            "Image:0",
+            ("Image:0",),
+            (1, 1, 1, 1, 1),
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["Channel:0:0"],
+        ),
+        # Test that we can support many dimensions if dims is provided
+        (
+            np.random.rand(1, 2, 3, 4, 5, 6, 7, 8),
+            "ABCDEFGH",
+            None,
+            "Image:0",
+            ("Image:0",),
+            (1, 3, 1, 1, 1),
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["Channel:0:0", "Channel:0:1", "Channel:0:2"],
+        ),
+        (
+            [
+                np.random.rand(1, 2, 3, 4, 5, 6, 7, 8),
+                da.random.random((1, 2, 3, 4, 5, 6, 7, 8)),
+            ],
+            "ABCDEFGH",
+            None,
+            "Image:0",
+            ("Image:0", "Image:1"),
+            (1, 3, 1, 1, 1),
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["Channel:0:0", "Channel:0:1", "Channel:0:2"],
+        ),
+        (
+            [
+                np.random.rand(1, 2, 3, 4, 5, 6, 7, 8),
+                da.random.random((2, 3, 4, 5, 6, 7, 8, 9)),
+            ],
+            "ABCDEFGH",
+            None,
+            "Image:1",
+            ("Image:0", "Image:1"),
+            (1, 4, 1, 1, 1),
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["Channel:1:0", "Channel:1:1", "Channel:1:2", "Channel:1:3"],
+        ),
+        # Test that without dims and with more than five dims, it raises an error
+        # Our guess dim order only support up to five dims
+        pytest.param(
+            np.random.rand(1, 2, 3, 4, 5, 6),
+            None,
+            None,
+            None,
+            None,
+            None,
+            None,
+            None,
+            marks=pytest.mark.raises(
+                exceptions=exceptions.InvalidDimensionOrderingError
+            ),
+        ),
+        pytest.param(
+            "hello world",
+            None,
+            None,
+            None,
+            None,
+            None,
+            None,
+            None,
+            marks=pytest.mark.raises(exceptions=exceptions.UnsupportedFileFormatError),
+        ),
+    ],
+)
+def test_aicsimage_from_array(
+    image: Union[types.MetaArrayLike, List[types.MetaArrayLike]],
+    dim_order: Optional[str],
+    channel_names: Optional[List[str]],
+    set_scene: str,
+    expected_scenes: Tuple[str, ...],
+    expected_shape: Tuple[int, ...],
+    expected_dims: str,
+    expected_channel_names: List[str],
+) -> None:
+    # Init
+    image_container = AICSImage(image, dim_order=dim_order, channel_names=channel_names)
 
     run_image_container_checks(
         image_container=image_container,
