+++ aicsimageio/tests/readers/test_ome_tiff_reader.py
@@ -1,14 +1,19 @@
 #!/usr/bin/env python
 # -*- coding: utf-8 -*-
 
-from typing import List, Tuple
+from typing import Any, Dict, List, Optional, Tuple, Type, Union
 
 import numpy as np
 import pytest
 from ome_types import OME
+from pathlib import Path
+from tempfile import TemporaryDirectory
+from distributed import LocalCluster, Client
+import time
 
-from aicsimageio import dimensions, exceptions
+from aicsimageio import AICSImage, dimensions, exceptions
 from aicsimageio.readers import OmeTiffReader
+from aicsimageio.readers.reader import Reader
 
 from ..conftest import LOCAL, get_resource_full_path, host
 from ..image_container_test_utils import (
@@ -470,3 +475,379 @@ def test_micromanager_ome_tiff_main_file() -> None:
     #     expected_channel_names=["Cy5", "DAPI", "FITC"],
     #     expected_physical_pixel_sizes=(1.75, 2.0, 2.0),
     # )
+
+
+@pytest.mark.parametrize(
+    "filename, "
+    "set_scene, "
+    "expected_scenes, "
+    "expected_shape, "
+    "expected_dtype, "
+    "expected_dims_order, "
+    "expected_channel_names, "
+    "expected_physical_pixel_sizes, "
+    "expected_metadata_type",
+    [
+        (
+            "s_1_t_1_c_10_z_1.ome.tiff",
+            "Image:0",
+            ("Image:0",),
+            (1, 10, 1, 1736, 1776),
+            np.uint16,
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            [f"C:{i}" for i in range(10)],  # This is the actual metadata
+            (None, None, None),
+            OME,
+        ),
+        (
+            # This is actually an OME-TIFF file
+            # Shows we don't just work off of extensions
+            # But the content of the file
+            "s_1_t_1_c_2_z_1_RGB.tiff",
+            "Image:0",
+            ("Image:0",),
+            (1, 2, 1, 32, 32, 3),
+            np.uint8,
+            dimensions.DEFAULT_DIMENSION_ORDER_WITH_SAMPLES,
+            ["Channel:0:0", "Channel:0:1"],
+            (None, None, None),
+            OME,
+        ),
+        (
+            "s_3_t_1_c_3_z_5.ome.tiff",
+            "Image:0",
+            ("Image:0", "Image:1", "Image:2"),
+            (1, 3, 5, 325, 475),
+            np.uint16,
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            ["EGFP", "TaRFP", "Bright"],
+            (1.0, 1.0833333333333333, 1.0833333333333333),
+            OME,
+        ),
+        (
+            "actk.ome.tiff",
+            "Image:0",
+            ("Image:0",),
+            (1, 6, 65, 233, 345),
+            np.float64,
+            dimensions.DEFAULT_DIMENSION_ORDER,
+            [
+                "nucleus_segmentation",
+                "membrane_segmentation",
+                "dna",
+                "membrane",
+                "structure",
+                "brightfield",
+            ],
+            (0.29, 0.29, 0.29),
+            OME,
+        ),
+    ],
+)
+def test_aicsimage(
+    filename: str,
+    set_scene: str,
+    expected_scenes: Tuple[str, ...],
+    expected_shape: Tuple[int, ...],
+    expected_dtype: np.dtype,
+    expected_dims_order: str,
+    expected_channel_names: List[str],
+    expected_physical_pixel_sizes: Tuple[float, float, float],
+    expected_metadata_type: Union[type, Tuple[Union[type, Tuple[Any, ...]], ...]],
+) -> None:
+    # Construct full filepath
+    uri = get_resource_full_path(filename, LOCAL)
+
+    # Run checks
+    run_image_file_checks(
+        ImageContainer=AICSImage,
+        image=uri,
+        set_scene=set_scene,
+        expected_scenes=expected_scenes,
+        expected_current_scene=set_scene,
+        expected_shape=expected_shape,
+        expected_dtype=expected_dtype,
+        expected_dims_order=expected_dims_order,
+        expected_channel_names=expected_channel_names,
+        expected_physical_pixel_sizes=expected_physical_pixel_sizes,
+        expected_metadata_type=expected_metadata_type,
+    )
+
+
+@pytest.mark.parametrize(
+    "filename, "
+    "first_scene_id, "
+    "first_scene_shape, "
+    "second_scene_id, "
+    "second_scene_shape",
+    [
+        (
+            "s_3_t_1_c_3_z_5.ome.tiff",
+            "Image:0",
+            (1, 3, 5, 325, 475),
+            "Image:1",
+            (1, 3, 5, 325, 475),
+        ),
+        (
+            "s_3_t_1_c_3_z_5.ome.tiff",
+            "Image:1",
+            (1, 3, 5, 325, 475),
+            "Image:2",
+            (1, 3, 5, 325, 475),
+        ),
+        pytest.param(
+            "s_1_t_1_c_1_z_1.ome.tiff",
+            "Image:1",
+            None,
+            None,
+            None,
+            marks=pytest.mark.raises(exception=IndexError),
+        ),
+        pytest.param(
+            "s_3_t_1_c_3_z_5.ome.tiff",
+            "Image:3",
+            None,
+            None,
+            None,
+            marks=pytest.mark.raises(exception=IndexError),
+        ),
+    ],
+)
+def test_multi_scene_aicsimage(
+    filename: str,
+    first_scene_id: str,
+    first_scene_shape: Tuple[int, ...],
+    second_scene_id: str,
+    second_scene_shape: Tuple[int, ...],
+) -> None:
+    # Construct full filepath
+    uri = get_resource_full_path(filename, LOCAL)
+
+    # Run checks
+    run_multi_scene_image_read_checks(
+        ImageContainer=AICSImage,
+        image=uri,
+        first_scene_id=first_scene_id,
+        first_scene_shape=first_scene_shape,
+        first_scene_dtype=np.dtype(np.uint16),
+        second_scene_id=second_scene_id,
+        second_scene_shape=second_scene_shape,
+        second_scene_dtype=np.dtype(np.uint16),
+    )
+
+
+@pytest.mark.parametrize(
+    "filename, set_scene, get_dims, get_specific_dims, expected_shape",
+    [
+        # Check normal volumetric data
+        (
+            "3d-cell-viewer.ome.tiff",
+            "Image:0",
+            "CZYX",
+            {"C": [0, 1, 2, 3]},
+            (4, 74, 1024, 1024),
+        ),
+    ],
+)
+@pytest.mark.parametrize("chunk_dims", ["YX", "ZYX"])
+@pytest.mark.parametrize("processes", [True, False])
+def test_parallel_read(
+    filename: str,
+    set_scene: str,
+    chunk_dims: str,
+    processes: bool,
+    get_dims: str,
+    get_specific_dims: Dict[str, Union[int, slice, range, Tuple[int, ...], List[int]]],
+    expected_shape: Tuple[int, ...],
+) -> None:
+    """
+    This test ensures that our produced dask array can be read in parallel.
+    """
+    # Construct full filepath
+    uri = get_resource_full_path(filename, LOCAL)
+
+    # Init image
+    img = AICSImage(uri, chunk_dims=chunk_dims)
+    img.set_scene(set_scene)
+
+    # Init cluster
+    cluster = LocalCluster(processes=processes)
+    client = Client(cluster)
+
+    # Select data
+    out = img.get_image_dask_data(get_dims, **get_specific_dims).compute()
+    assert out.shape == expected_shape
+
+    # Shutdown and then safety measure timeout
+    cluster.close()
+    client.close()
+    time.sleep(5)
+
+
+@pytest.mark.parametrize(
+    "filename, "
+    "reconstruct_mosaic, "
+    "set_scene, "
+    "expected_shape, "
+    "expected_mosaic_tile_dims, "
+    "specific_tile_index",
+    [
+        pytest.param(
+            "actk.ome.tiff",
+            True,
+            "Image:0",
+            (1, 6, 65, 233, 345),
+            None,
+            0,
+            # AttributeError raises not because of error in rollback
+            # but because cannot access Y or X from
+            # None return from `mosaic_tile_dims` because
+            # image is not a mosaic tiled image
+            marks=pytest.mark.raises(exception=AttributeError),
+        ),
+    ],
+)
+def test_mosaic_passthrough(
+    filename: str,
+    reconstruct_mosaic: bool,
+    set_scene: str,
+    expected_shape: Tuple[int, ...],
+    expected_mosaic_tile_dims: Tuple[int, int],
+    specific_tile_index: int,
+) -> None:
+    # Construct full filepath
+    uri = get_resource_full_path(filename, LOCAL)
+
+    # Init
+    img = AICSImage(uri, reconstruct_mosaic=reconstruct_mosaic)
+    img.set_scene(set_scene)
+
+    # Assert basics
+    assert img.shape == expected_shape
+    assert img.mosaic_tile_dims.Y == expected_mosaic_tile_dims[0]  # type: ignore
+    assert img.mosaic_tile_dims.X == expected_mosaic_tile_dims[1]  # type: ignore
+
+    # Ensure that regardless of stitched or not, we can get tile position
+    img.get_mosaic_tile_position(specific_tile_index)
+
+
+@pytest.mark.parametrize(
+    "filename, expected_reader",
+    [
+        (
+            "s_1_t_1_c_1_z_1.ome.tiff",
+            OmeTiffReader,
+        ),
+        (  # Multiscene tiff
+            "variable_scene_shape_first_scene_pyramid.ome.tiff",
+            OmeTiffReader,
+        ),
+    ],
+)
+def test_selected_tiff_reader(
+    filename: str,
+    expected_reader: Type[Reader],
+) -> None:
+    # Construct full filepath
+    uri = get_resource_full_path(filename, LOCAL)
+
+    # Init
+    img = AICSImage(uri)
+
+    # Assert basics
+    assert isinstance(img.reader, expected_reader)
+
+
+@pytest.mark.parametrize(
+    "filename, set_reader, extra_kwargs, expected_dims, expected_shape",
+    [
+        # Test incompatible reader
+        pytest.param(
+            "s_1_t_1_c_1_z_1.tiff",
+            OmeTiffReader,
+            {},
+            None,
+            None,
+            marks=pytest.mark.raises(exceptions=exceptions.UnsupportedFileFormatError),
+        ),
+    ],
+)
+def test_set_reader(
+    filename: str,
+    set_reader: Type[Reader],
+    extra_kwargs: Dict[str, Any],
+    expected_dims: str,
+    expected_shape: Tuple[int, ...],
+) -> None:
+    # Construct full filepath
+    uri = get_resource_full_path(filename, LOCAL)
+
+    # Init
+    img = AICSImage(uri, reader=set_reader, **extra_kwargs)
+
+    # Assert basics
+    assert img.dims.order == expected_dims
+    assert img.shape == expected_shape
+
+
+@pytest.mark.parametrize(
+    "filename, select_scenes",
+    [
+        ("s_3_t_1_c_3_z_5.ome.tiff", None),
+        ("s_3_t_1_c_3_z_5.ome.tiff", ["Image:2", "Image:1"]),
+    ],
+)
+def test_roundtrip_save_all_scenes(
+    filename: str, select_scenes: Optional[List[str]]
+) -> None:
+    # Construct full filepath
+    uri = get_resource_full_path(filename, LOCAL)
+
+    # Read initial
+    original = AICSImage(uri)
+
+    # Save to temp and compare
+    with TemporaryDirectory() as tmpdir:
+        save_path = Path(tmpdir) / f"converted-{filename}.ome.tiff"
+        original.save(save_path, select_scenes=select_scenes)
+
+        # Re-read
+        result = AICSImage(save_path)
+
+        # Compare all scenes
+        # They may not have the same scene ids as some readers use scene names as the
+        # id, see LifReader for example
+        if select_scenes is None:
+            select_original_scenes = list(original.scenes)
+        else:
+            select_original_scenes = select_scenes
+
+        assert len(select_original_scenes) == len(result.scenes)
+        for original_scene_id, result_scene_id in zip(
+            select_original_scenes, result.scenes
+        ):
+            # Compare
+            original.set_scene(original_scene_id)
+            result.set_scene(result_scene_id)
+
+            np.testing.assert_array_equal(original.data, result.data)
+            assert original.dims.order == result.dims.order
+            assert original.channel_names == result.channel_names
+
+
+@pytest.mark.parametrize(
+    "filename, expected_shape",
+    [
+        ("s_1_t_1_c_10_z_1.ome.tiff", (1, 10, 1, 1736, 1776)),
+    ],
+)
+def test_no_scene_prop_access(
+    filename: str,
+    expected_shape: Tuple[int, ...],
+) -> None:
+    # Construct full filepath
+    uri = get_resource_full_path(filename, LOCAL)
+
+    # Construct image and check no scene call with property access
+    img = AICSImage(uri)
+    assert img.shape == expected_shape
