+++ aicsimageio/tests/writers/test_two_d_writer.py
@@ -8,7 +8,7 @@ import pytest
 
 from aicsimageio import exceptions
 from aicsimageio.readers.default_reader import DefaultReader
-from aicsimageio.writers.two_d_writer import TwoDWriter
+from aicsimageio.writers.extra_writers.two_d_writer import TwoDWriter
 
 from ..conftest import LOCAL, array_constructor, get_resource_write_full_path
