+++ aicsimageio/writers/extra_writers/timeseries_writer.py
@@ -8,15 +8,15 @@ import numpy as np
 from fsspec.implementations.local import LocalFileSystem
 from imageio import get_writer
 
-from .. import types
-from ..dimensions import DimensionNames
-from ..exceptions import InvalidDimensionOrderingError, UnexpectedShapeError
-from ..transforms import reshape_data
-from ..utils import io_utils
-from .writer import Writer
+from ... import types
+from ...dimensions import DimensionNames
+from ...exceptions import InvalidDimensionOrderingError, UnexpectedShapeError
+from ...transforms import reshape_data
+from ...utils import io_utils
+from ..writer import Writer
 
 try:
-    from ..readers.default_reader import DefaultReader
+    from ...readers.default_reader import DefaultReader
 
 except ImportError:
     raise ImportError(
