+++ setup.py
@@ -36,7 +36,7 @@ format_libs: Dict[str, List[str]] = {
     ],
     "nd2": ["nd2[legacy]>=0.2.0"],
     "dv": ["mrc>=0.2.0"],
-    "bfio": ["bfio>=2.3.0"],
+    "bfio": ["bfio>=2.3.0", "tifffile<2022.4.22"],
     # "czi": ["aicspylibczi>=3.0.5"],  # excluded for licensing reasons
     # "bioformats": ["bioformats_jar"],  # excluded for licensing reasons
     # "lif": ["readlif>=0.6.4"],  # excluded for licensing reasons
@@ -52,7 +52,6 @@ setup_requirements = [
 ]
 
 test_requirements = [
-    *all_formats,
     "codecov>=2.1.4",
     "distributed>=2021.4.1",
     "docutils>=0.10,<0.16",
@@ -63,10 +62,6 @@ test_requirements = [
     "quilt3",  # no pin to avoid pip cycling (boto is really hard to manage)
     "s3fs[boto3]>=0.4.2",
     "tox>=3.15.2",
-    "bioformats_jar",  # to test bioformats
-    "bfio>=2.3.0",
-    "readlif>=0.6.4",  # to test lif
-    "aicspylibczi>=3.0.5",  # to test czi
 ]
 
 dev_requirements = [
@@ -88,6 +83,12 @@ dev_requirements = [
     "sphinx_rtd_theme>=0.5.1",
     "twine>=3.1.1",
     "wheel>=0.34.2",
+    # reader deps
+    *all_formats,
+    "bioformats_jar",  # to test bioformats
+    "bfio>=2.3.0",
+    "readlif>=0.6.4",  # to test lif
+    "aicspylibczi>=3.0.5",  # to test czi
 ]
 
 benchmark_requirements = [
@@ -103,7 +104,7 @@ requirements = [
     "numpy>=1.16,<2",
     "ome-types>=0.2",
     "wrapt>=1.12",
-    "tifffile>=2022.4.26",
+    "tifffile>=2021.8.30",
     "xarray>=0.16.1",
     "xmlschema",  # no pin because it's pulled in from OME types
     "zarr>=2.6,<3",
