+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py38, py39, lint
+envlist = py38, py39, bioformats, czi, dv, lif, nd2, bfio, lint
 skip_missing_interpreters = true
 toxworkdir={env:TOX_WORK_DIR:.tox}
 
@@ -7,6 +7,88 @@ toxworkdir={env:TOX_WORK_DIR:.tox}
 skip_install = true
 deps = pre-commit
 commands = pre-commit run --all-files --show-diff-on-failure
+
+[testenv:bioformats]
+passenv =
+    AWS_*
+    JAVA_HOME
+    CI
+setenv =
+    PYTHONPATH = {toxinidir}
+deps = bioformats_jar
+commands =
+    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/extra_readers/test_bioformats_reader.py {posargs}
+
+[testenv:czi]
+passenv =
+    AWS_*
+    JAVA_HOME
+    CI
+setenv =
+    PYTHONPATH = {toxinidir}
+deps = aicspylibczi>=3.0.5
+commands =
+    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/extra_readers/test_czi_reader.py {posargs}
+
+[testenv:default]
+passenv =
+    AWS_*
+    JAVA_HOME
+    CI
+setenv =
+    PYTHONPATH = {toxinidir}
+extras =
+    imageio
+commands =
+    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/extra_readers/test_default_reader.py {posargs}
+
+[testenv:dv]
+passenv =
+    AWS_*
+    JAVA_HOME
+    CI
+setenv =
+    PYTHONPATH = {toxinidir}
+extras =
+    dv
+commands =
+    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/extra_readers/test_dv_reader.py {posargs}
+
+[testenv:lif]
+passenv =
+    AWS_*
+    JAVA_HOME
+    CI
+setenv =
+    PYTHONPATH = {toxinidir}
+deps = readlif>=0.6.4
+commands =
+    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/extra_readers/test_lif_reader.py {posargs}
+
+[testenv:nd2]
+passenv =
+    AWS_*
+    JAVA_HOME
+    CI
+setenv =
+    PYTHONPATH = {toxinidir}
+extras =
+    nd2
+commands =
+    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/extra_readers/test_nd2_reader.py {posargs}
+
+[testenv:bfio]
+passenv =
+    AWS_*
+    JAVA_HOME
+    CI
+setenv =
+    PYTHONPATH = {toxinidir}
+extras =
+    bfio
+    test
+commands =
+    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/readers/extra_readers/test_ome_tiled_tiff_reader.py {posargs}
 
 [testenv]
 passenv =
@@ -18,4 +100,4 @@ setenv =
 extras =
     test
 commands =
-    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/ {posargs}
+    pytest --basetemp={envtmpdir} --cov-report xml --cov-report html --cov=aicsimageio aicsimageio/tests/ --ignore-glob="**/extra_readers/*" --ignore-glob="**/extra_writers/*" {posargs}
